+++ src/main/java/chess/Applictaion.java
@@ -1,13 +1,11 @@
 package chess;
 
 import chess.controller.ChessController;
-import chess.view.InputView;
-import chess.view.OutputView;
 
 public class Applictaion {
     
     public static void main(String[] args) {
-        ChessController chessController = new ChessController(new InputView(), new OutputView());
+        ChessController chessController = new ChessController();
         chessController.execute();
     }
 }
