+++ src/main/java/chess/controller/ChessController.java
@@ -1,22 +1,18 @@
 package chess.controller;
 
-import chess.CommandLine;
+import chess.controller.dto.BoardDto;
 import chess.domain.ChessGame;
+import chess.domain.CommandLine;
 import chess.view.InputView;
 import chess.view.OutputView;
 
 public class ChessController {
-
-    private final InputView inputView;
-    private final OutputView outputView;
     
-    public ChessController(final InputView inputView, final OutputView outputView) {
-        this.inputView = inputView;
-        this.outputView = outputView;
+    public ChessController() {
     }
     
     public void execute() {
-        this.outputView.printGameStartMessage();
+        OutputView.printGameStartMessage();
         ChessGame chessGame = new ChessGame();
         while (!chessGame.isGameEnd()) {
             this.runGame(chessGame);
@@ -26,15 +22,15 @@ public class ChessController {
     private void runGame(final ChessGame chessGame) {
         try {
             CommandLine commandLine = this.getCommandLine();
-            this.handleCommandLine(chessGame, commandLine);
-            this.outputView.printBoard(BoardDto.create(chessGame.getBoard()));
+            handleCommandLine(chessGame, commandLine);
+            OutputView.printBoard(BoardDto.create(chessGame.getBoard()));
         } catch (Exception e) {
-            this.outputView.printError(e.getMessage());
+            OutputView.printError(e.getMessage());
         }
     }
     
     private CommandLine getCommandLine() {
-        return new CommandLine(this.inputView.readCommand());
+        return new CommandLine(InputView.readCommand());
     }
     
     private void handleCommandLine(final ChessGame chessGame, final CommandLine commandLine) {
