+++ src/main/java/chess/domain/GameStatus.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.domain;
 
 public enum GameStatus {
     START,
