+++ src/main/java/chess/view/InputView.java
@@ -10,7 +10,7 @@ public class InputView {
     private static final Scanner scanner = new Scanner(System.in);
     private static final String DELIMITER = " ";
     
-    public List<String> readCommand() {
+    public static List<String> readCommand() {
         String input = scanner.nextLine();
         return Arrays.stream(input.split(DELIMITER)).collect(Collectors.toList());
     }
